Multi-mesh levels
Warning: this breaks Level 1

But level 2 now demonstrates the possibility of mesh composition.

Basically, I want to be able to tag meshes as static parts of the terrain.

I'm doing this by changing up the world class to instead of treating the
HterrainFullmesh as magically the entire terrain mesh, adding a component
to various meshes which will cause them to register as part of the
terrain mesh.

At level start, the world will compose all of these together into an
ArrayMesh.

Surprisingly, as you can see from the debugger, this actually works.
The only catch is that it does not account for the transforms of the meshes,
so we'll need to apply those xforms (see the TODO in World.gd

Further experimentation is required to see how far we can go with this
(ramps? Bridges? etc)